refactor: Extract listing of all databases into iox_object_store
